+++ app/services/import/hash/tarif_importer.rb
@@ -6,9 +6,9 @@ module Import
       attr_reader :organisation
 
       use_attributes(*%w[associated_types label_i18n ordinal prefill_usage_method price_per_unit tarif_group
-                         pin type unit_i18n vat_category_id valid_from valid_until
+                         pin type unit_i18n vat_category_id valid_from valid_until mode
                          accounting_account_nr accounting_cost_center_nr
-                         minimum_price_per_night minimum_price_total minimum_usage_per_night minimum_usage_total])
+                         minimum minimum_mode included_units included_units_mode])
 
       def initialize(organisation, **)
         super(**)