+++ spec/models/tarif_spec.rb
@@ -10,11 +10,11 @@
 #  associated_types                  :integer          default(0), not null
 #  discarded_at                      :datetime
 #  enabling_conditions               :jsonb
+#  included_units                    :decimal(, )
+#  included_units_mode               :integer          default(0)
 #  label_i18n                        :jsonb
-#  minimum_price_per_night           :decimal(, )
-#  minimum_price_total               :decimal(, )
-#  minimum_usage_per_night           :decimal(, )
-#  minimum_usage_total               :decimal(, )
+#  minimum                           :decimal(, )
+#  minimum_mode                      :integer          default(0)
 #  mode                              :integer
 #  ordinal                           :integer
 #  pin                               :boolean          default(TRUE)
@@ -50,41 +50,5 @@ RSpec.describe Tarif do
     subject(:price) { usage.price }
 
     it { is_expected.to eq(70.0) }
-  end
-
-  describe '#minimum_prices' do
-    subject(:minimum_prices) { usage.minimum_prices }
-
-    before do
-      tarif.update({
-                     minimum_usage_per_night: 24,
-                     minimum_usage_total: 71,
-                     minimum_price_per_night: 210,
-                     minimum_price_total: 610
-                   })
-    end
-
-    it 'lists all minimum prices' do
-      expect(minimum_prices).to eq({
-                                     minimum_usage_per_night: 24 * 7 * 10,
-                                     minimum_usage_total: 71 * 10,
-                                     minimum_price_per_night: 210 * 7,
-                                     minimum_price_total: 610
-                                   })
-    end
-  end
-
-  describe '#minimum_price' do
-    before do
-      tarif.update({
-                     minimum_usage_per_night: 24,
-                     minimum_usage_total: 71,
-                     minimum_price_per_night: 210,
-                     minimum_price_total: 610
-                   })
-    end
-
-    it { expect(usage.minimum_price).to eq(24 * 7 * 10) }
-    it { expect(usage.critical_minimum).to eq(:minimum_usage_per_night) }
   end
 end